Renaming file, updaing to Object.keys.find pattern
