+++ src/omnibar/omnibar-user-session-expiration.spec.ts
@@ -1,4 +1,4 @@
-import { BBAuthDomain } from '../auth/auth-get-domain';
+import { BBAuthDomain } from '../auth/auth-domain';
 import { BBCsrfXhr } from '../shared/csrf-xhr';
 import { BBOmnibarUserSessionExpiration } from './omnibar-user-session-expiration';
